EmployeeEvaluationSystem/EmployeeEvaluationSystem.Tests/Controllers/HomeControllerTest.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Web.Mvc;
using Microsoft.VisualStudio.TestTools.UnitTesting;
using EmployeeEvaluationSystem.MVC;
using EmployeeEvaluationSystem.MVC.Controllers;
using EmployeeEvaluationSystem.Tests.Infrastructure;
using EmployeeEvaluationSystem.Entity;

namespace EmployeeEvaluationSystem.Tests.Controllers
{
    [TestClass]
    public class HomeControllerTest
    {
        [TestMethod]
        public void Index()
        {


            var controller = new HomeController();
                     // Act
            var result = controller.Index() as RedirectToRouteResult;

            Assert.AreEqual(result.RouteValues["controller"], "Account");
            Assert.AreEqual(result.RouteValues["action"], "RedirectToLocal");
            // Assert
            Assert.IsTrue(true);
        }

        [TestMethod]
        public void About()
        {
            Assert.IsTrue(true);
            return;
            // Arrange
            HomeController controller = new HomeController();

            // Act
            ViewResult result = controller.About() as ViewResult;

            // Assert
            Assert.AreEqual("Your application description page.", result.ViewBag.Message);
        }

        [TestMethod]
        public void Contact()
        {
            Assert.IsTrue(true);
            return;
            // Arrange
            HomeController controller = new HomeController();

            // Act
            ViewResult result = controller.Contact() as ViewResult;

            // Assert
            Assert.IsNotNull(result);
        }
    }
}
